Asymptotic Variances of the MDE

Introduction

In the multiscale overdamped Langevin drift parameter estimation problem, cf. section 4 of the article XXX, the MDE is asymptotically normal under the true parameter $\vartheta_0 \in \Theta_0$ as $\epsilon \rightarrow 0$ with

\[\begin{aligned} \sqrt{T_\epsilon} \left(\hat{\vartheta}_{T_\epsilon}(X_\epsilon) - \vartheta_0 \right) \overset{\mathcal{D}_{\psi(\vartheta_0)}}{\longrightarrow} J(\vartheta_0)^{-1} \mathcal{N}_1(0, \tau^2(\vartheta_0)), \quad \text{as } \epsilon \rightarrow 0. \end{aligned}\]

The following functions calculate the asymptotic variance figuring above.

Functions

MDE_project.Σ_∞_QdP — Function

Σ_∞_QdP(ϑ, Σ)

Return the asymptotic variance of the MDE in the multiscale overdamped Langevin drift parameter estimation problem with a quadratic potential, drift parameter ϑ, and diffusion parameter Σ.

The asymptotic variance is given by

\[\begin{aligned} \frac{\tau^2(\vartheta, \Sigma)}{J(\vartheta, \Sigma)^2} = \frac{2\Sigma}{J(\vartheta, \Sigma)^{2}} \int_\R |\Phi'(x)|^2 \mu(x, \vartheta, \Sigma, V) \, dx, \end{aligned}\]

where

\[\begin{aligned} J(\vartheta, \Sigma) = \| \partial_\vartheta \mathscr{C}_{\vartheta} \|_{L^2(\varphi)}, \quad \Phi'(x) = \frac{1}{\Sigma \mu(x, \vartheta, \Sigma, V)} \int_{-\infty}^x h(z) \mu(x, \vartheta, \Sigma, V) \, dz, \quad x \in \R. \end{aligned}\]

Here $\mu(x, \vartheta, \Sigma, V)$ corresponds to μ with $V(x) = x^2/2$. In the case of such a quadratic potential $V$ it holds

\[\begin{aligned} J(\vartheta, \Sigma) = \frac{3}{4 \beta} \left( \frac{\Sigma}{\vartheta^2} \right)^2 \sigma_2^5, \quad h(z) = \frac{\Sigma}{2 \vartheta^2 \beta} \left[ \sigma_1^3 (1 - \sigma_1^2 z^2) \exp \left( -\frac{\sigma_1^2 z^2}{2} \right) - \sigma_2^3 \right], \quad z \in \R, \end{aligned}\]

with

\[\begin{aligned} \sigma_1^2 := \frac{\beta^2 \vartheta}{\vartheta + \Sigma\beta^2 }, \quad \sigma_2^2 := \frac{\beta^2 \vartheta}{\vartheta + 2 \Sigma\beta^2 }. \end{aligned}\]

Warning

The running times for a single evaluation can take some time due to the involved integrations, but it usually needs to be computed only once.

Arguments

ϑ::Real: positive drift coefficient $\vartheta$.

Σ::Real: positive diffusion coefficient $\Sigma$.

Examples

julia> Σ_∞_QdP(1.2, 0.6)

See also μ, Σ_∞_QrP.

MDE_project.Σ_∞_QrP — Function

Σ_∞_QrP(ϑ, Σ)

Return the asymptotic variance of the MDE in the multiscale overdamped Langevin drift parameter estimation problem with a quartic potential, drift parameter ϑ, and diffusion parameter Σ.

The asymptotic variance is given by

\[\begin{aligned} \frac{\tau^2(\vartheta, \Sigma)}{J(\vartheta, \Sigma)^2} = \frac{2\Sigma}{J(\vartheta, \Sigma)^{2}} \int_\R |\Phi'(x)|^2 \mu(x, \vartheta, \Sigma, V) \, dx, \end{aligned}\]

where

\[\begin{aligned} J(\vartheta, \Sigma) = \| \partial_\vartheta \mathscr{C}_{\vartheta} \|_{L^2(\varphi)}, \quad \Phi'(x) = \frac{1}{\Sigma \mu(x, \vartheta, \Sigma, V)} \int_{-\infty}^x h(z) \mu(x, \vartheta, \Sigma, V) \, dz, \quad x \in \R. \end{aligned}\]

Here $\mu(x, \vartheta, \Sigma, V)$ corresponds to μ with $V(x) = x^4/4-x^2/2$. In this case it holds

\[\begin{aligned} &J(\vartheta, \Sigma) = \int_\R (\partial_\vartheta \mu(\vartheta, \Sigma, V) \ast k_\beta)(x) \, \partial_\vartheta \mu(x, \vartheta, \Sigma, V) \, dx, \\[0.5cm] &h(z) = (\partial_\vartheta \mu(\vartheta, \Sigma, V) \ast k_\beta)(z) - \int_\R (\partial_\vartheta \mu(\vartheta, \Sigma, V) \ast k_\beta)(x) \, \mu(x, \vartheta, \Sigma, V) \, dx. \end{aligned}\]

Warning

The running times for a single evaluation can take some time due to the involved integrations, but it usually needs to be computed only once.

Arguments

ϑ::Real: positive drift coefficient $\vartheta$.

Σ::Real: positive diffusion coefficient $\Sigma$.

Examples

julia> Σ_∞_QrP(1.2, 0.6)

See also μ, k, Σ_∞_QdP.

Index

MDE_project.Σ_∞_QdP

MDE_project.Σ_∞_QrP