Cost Functionals for the MDE

In its most general form, the cost functional of the MDE is given by

\[\begin{aligned} \Delta_T(\vartheta, X_\epsilon) := \int_{\R^d} \left| \frac1T \int_0^T \exp\left(i u^\top X_\epsilon(t)\right) \, dt - \int_{\R^d} \exp\left(i u^\top x\right) \mu(\vartheta, x) \, dx \right|^2 \varphi(u) \, du. \end{aligned}\]

Choosing a centered Gaussian density with covariance matrix $\beta^2 I_d, \beta > 0$, for $\varphi$ reduces the above formula to

\[\begin{aligned} \Delta_T(\vartheta, X_\epsilon) = &\frac{1}{T^2} \int_0^T \int_0^T k(X_\epsilon(t)-X_\epsilon(s)) \, dt \, ds \\[0.25cm] &- \frac{2}{T} \int_0^T (\mu(\vartheta) \ast k)(X_\epsilon(t)) \, dt + \int_{\R^d} (\mu(\vartheta) \ast k)(x) \mu(\vartheta, x) \, dx, \end{aligned}\]

where $k(x) :=\exp(-\beta^2 |x|_2^2/2)$ is the characteristic function of $\varphi$ and $\ast$ denotes the convolution operator on $\R^d$. Furthermore, if $\mu(\vartheta)$ is the density of a centered multivariate normal distribution with a positive definite covariance matrix $M(\vartheta) \in \R^{d \times d}$, then

\[\begin{aligned} \Delta_T(\vartheta, X_\epsilon) &= - \frac{2}{T \sqrt{\det(I_d + \beta^2 M(\vartheta)) }} \int_0^T \exp\left( -\frac{\beta^2}{2} X_\epsilon(t)^\top \left( I_d + \beta^2 M(\vartheta) \right)^{-1} X_\epsilon(t) \right) \, dt \\[0.25cm] &+ \frac{1}{\sqrt{\det(I_d + 2 \beta^2 M(\vartheta))}}. \end{aligned}\]

Cost Functionals

MDE_project.k — Function

k(x, β)

Return β and function value of the characteristic function of a centered Gaussian density with standard deviation β at the point x as a tuple.

This characteristic function is given by

\[\begin{aligned} k_\beta(x) = \exp\left( -\frac{\beta^2 x^2}{2} \right), \quad x \in \R. \end{aligned}\]

It is used in the definition of the MDE and thoroughly outlined in the main manuscript in the numerics section.

Arguments

x::Real: argument $x$ at which to evaluate the function.

β::Real: positive number $\beta$.

MDE_project.Δ — Function

Δ(data, ϑ, Σ, V)

Compute cost functional for given data and parameter values ϑ and Σ.

A properly discretized version of the cost functional, given by

\[\begin{aligned} \Delta_T(X_\epsilon, \vartheta, \Sigma, V) = - \frac{2}{T} \int_0^T (\mu(\vartheta, \Sigma, V) \ast k_\beta)(X_\epsilon(t)) \, dt + \int_{\R} (\mu(\vartheta, \Sigma, V) \ast k_\beta)(x) \mu(\vartheta, \Sigma, x) \, dx, \end{aligned}\]

is implemented and evaluated via multithreading. Here, $X_ϵ$ is a one-dimensional time series of length $T$, obtained from a multiscale SDE, $\mu$ is the invariant density of the homogenized limit SDE, $k_\beta$ refers to k, and $\ast$ is the convolution operator on $\R$. See the main manuscript for details on this functional. It is the core object of the MDE.

Warning

The computational cost of this function is quite high due to the integration of the convolutions, so if the data is finely discretized, then the running times for a single evalutation are relatively long. Remember that, in this case of a non-quadratic potential, further simplifications of the above formula are not known thus far.

Note

When comparing the above formula with the formula from the main manuscript, then one notices that the double integral term is missing above. This is on purpose because the double integral does not depend on any parameters with respect to which we will optimize.

Arguments

data::Vector{Real}: one-dimensional time series $X_ϵ$.

ϑ::Real: positive drift coefficient $\vartheta$.

Σ::Real: positive diffusion coefficient $\Sigma$.

V: defining potential function $V$ for the invariant density.

Examples

$ julia --threads 10 --project=. # start julia with 10 threads and activate project

julia> Threads.nthreads()
julia> using MDE_project
julia> data = Langevin_ϵ(1.0, func_config=NLDO(), α=2.0, σ=1.0, ϵ=0.1, T=100)[1]
julia> Δ(data, 1, 1, NLDO()[1])

See also Δ_Gaussian1D, Δ_Gaussian2D.

MDE_project.Δ_Gaussian1D — Function

Δ_Gaussian1D(data, ϑ, Σ)

Compute cost functional for given one-dimensonal data and parameter values ϑ and Σ in the case where the invariant density of the homogenized limit SDE is centered Gaussian.

A properly discretized version of the cost functional, given by

\[\begin{aligned} \Delta_T(X_\epsilon, \vartheta, \Sigma, V) = -\frac{2}{T \sqrt{1 + \beta^2 \frac{\Sigma}{\vartheta}} } \int_0^T \exp\left( -\frac{\beta^2 X_\epsilon(t)^2}{2 (1 + \beta^2 \frac{\Sigma}{\vartheta})} \right) \, dt + \frac{1}{\sqrt{1 + 2 \beta^2 \frac{\Sigma}{\vartheta}}}, \end{aligned}\]

is implemented. Here, $X_ϵ$ is a one-dimensional time series of length $T$, obtained from a multiscale SDE, and $\beta$ comes from k. The potential is here $V(x) = x^2/2$. See the main manuscript for details on this functional. It is the core object of the MDE.

Note

The evaluation is, compared to Δ, extremely fast, even for finely discretized data, which signifies the utility of choosing a centered Gaussian weight k in this Gaussian case.

Arguments

data::Vector{Real}: one-dimensional time series $X_ϵ$.

ϑ::Real: positive drift coefficient $\vartheta$.

Σ::Real: positive diffusion coefficient $\Sigma$.

Examples

julia> using MDE_project
julia> data = Langevin_ϵ(1.0, func_config=LDO(), α=2.0, σ=1.0, ϵ=0.1, T=100)[1]
julia> Δ_Gaussian1D(data, 1, 1)

See also Δ, Δ_Gaussian2D.

MDE_project.Δ_Gaussian2D — Function

Δ_Gaussian2D(data, ϑ, Σ)

Compute cost functional for given two-dimensonal data and parameter values ϑ and Σ in the case where the invariant density of the homogenized limit SDE is centered Gaussian.

A properly discretized version of the cost functional, given by

\[\begin{aligned} \Delta_T(X_\epsilon, \vartheta, \Sigma) &= - \frac{2}{T \sqrt{\det(I_d + \beta^2 M(\vartheta, \Sigma)) }} \int_0^T \exp\left( -\frac{\beta^2}{2} X_\epsilon(t)^\top \left( I_d + \beta^2 M(\vartheta, \Sigma) \right)^{-1} X_\epsilon(t) \right) \, dt \\[0.25cm] &+ \frac{1}{\sqrt{\det(I_d + 2 \beta^2 M(\vartheta, \Sigma))}}, \end{aligned}\]

is implemented. Here, $X_ϵ$ is a two-dimensional time series of length $T$, obtained from a multiscale SDE, $M(\vartheta, \Sigma) \in \R^{2 \times 2}$ is a matrix depending on $\vartheta$ and $\Sigma$ and is given by the covariance matrix of the invariant Gaussian density, and $\beta$ comes from k. See the main manuscript for details on this functional and the appearing quantities. It is the core object of the MDE.

Note

The evaluation is, compared to Δ, extremely fast, even for finely discretized two-dimensional data, which signifies the utility of choosing a two-dimensional centered Gaussian weight k in this Gaussian case.

Arguments

data::Vector{Real}: two-dimensional time series $X_ϵ$.

ϑ::Array{Real}: positive definite drift matrix $\vartheta \in \mathbb{R}^{2 \times 2}$.

Σ::Array{Real}: positive definite diffusion matrix $\Sigma \in \mathbb{R}^{2 \times 2}$.

Examples

julia> using MDE_project
julia> M=[4 2;2 3]
julia> σ = 5.0  
julia> data = Langevin_ϵ_2D([-5.0, -5.0], func_config=(x-> cos(x), x -> 1/2*cos(x)), M=M, σ=σ, ϵ=0.1, T=100.0)[1]
julia> CorrK = [K(x-> cos(x), σ) 0 ; 0 K(x -> 1/2*cos(x), σ)]
julia> ϑ = CorrK*M
julia> Σ = σ*CorrK
julia> Δ_Gaussian2D(data, ϑ, Σ)

See also Δ, Δ_Gaussian1D.

Gradients of Cost Functionals

MDE_project.Δ_grad_ϑ — Function

Δ_grad_ϑ(data, ϑ, Σ, V)

Compute gradient of cost functional Δ with respect to ϑ for given data and parameter values ϑ and Σ and a potential V.

A properly discretized version of the gradient of the cost functional, given by

\[\begin{aligned} \partial_\vartheta \Delta_T(X_\epsilon, \vartheta, \Sigma, V) = -\frac{2}{T} \int_0^T (\partial_\vartheta \mu(\vartheta, \Sigma, V) \ast k_\beta)(X_\epsilon(t)) \, dt + 2 \int_{\R} (\mu(\vartheta, \Sigma, V) \ast k_\beta)(x) \partial_\vartheta \mu(\vartheta, \Sigma, x) \, dx, \end{aligned}\]

is implemented and evaluated via multithreading. Here, $X_ϵ$ is a one-dimensional time series of length $T$, obtained from a multiscale SDE, $\mu$ is the invariant density of the homogenized limit SDE corresponding to μ, $k_\beta$ refers to k, and $\ast$ is the convolution operator on $\R$.

Warning

The computational cost of this function is quite high due to the integration of the convolutions, so if the data is finely discretized, then the running times for a single evalutation are relatively long. Remember that, in this case of a non-quadratic potential, further simplifications of the above formula are not known thus far.

Arguments

data::Vector{Real}: one-dimensional time series $X_ϵ$.

ϑ::Real: drift coefficient $\vartheta$.

Σ::Real: positive diffusion coefficient $\Sigma$.

V: defining potential function $V$ for the invariant density.

Examples

$ julia --threads 10 --project=. # start julia with 10 threads and activate project

julia> Threads.nthreads()
julia> using MDE_project
julia> data = Langevin_ϵ(1.0, func_config=NLDO(), α=2.0, σ=1.0, ϵ=0.1, T=100)[1]
julia> Δ_grad_ϑ(data, 1, 1, NLDO()[1])

See also Δ.

MDE_project.Δ_grad_Σ — Function

Δ_grad_Σ(data, ϑ, Σ, V)

Compute gradient of cost functional Δ with respect to Σ for given data and parameter values ϑ and Σ and a potential V.

A properly discretized version of the gradient of the cost functional, given by

\[\begin{aligned} \partial_\Sigma \Delta_T(X_\epsilon, \vartheta, \Sigma, V) = -\frac{2}{T} \int_0^T (\partial_\Sigma \mu(\vartheta, \Sigma, V) \ast k_\beta)(X_\epsilon(t)) \, dt + 2 \int_{\R} (\mu(\vartheta, \Sigma, V) \ast k_\beta)(x) \partial_\Sigma \mu(\vartheta, \Sigma, x) \, dx, \end{aligned}\]

is implemented and evaluated via multithreading. Here, $X_ϵ$ is a one-dimensional time series of length $T$, obtained from a multiscale SDE, $\mu$ is the invariant density of the homogenized limit SDE corresponding to μ, $k_\beta$ refers to k, and $\ast$ is the convolution operator on $\R$.

Warning

The computational cost of this function is quite high due to the integration of the convolutions, so if the data is finely discretized, then the running times for a single evalutation are relatively long. Remember that, in this case of a non-quadratic potential, further simplifications of the above formula are not known thus far.

Arguments

data::Vector{Real}: one-dimensional time series $X_ϵ$.

ϑ::Real: drift coefficient $\vartheta$.

Σ::Real: positive diffusion coefficient $\Sigma$.

V: defining potential function $V$ for the invariant density.

Examples

$ julia --threads 10 --project=. # start julia with 10 threads and activate project

julia> Threads.nthreads()
julia> using MDE_project
julia> data = Langevin_ϵ(1.0, func_config=NLDO(), α=2.0, σ=1.0, ϵ=0.1, T=100)[1]
julia> Δ_grad_Σ(data, 1, 1, NLDO()[1])

See also Δ.

MDE_project.Δ_Gaussian1D_grad — Function

Δ_Gaussian1D_grad(data, ϑ, Σ)

Compute gradient of cost functional Δ_Gaussian1D for given one-dimensonal data and parameter values ϑ and Σ.

A properly discretized version of the gradient of the cost functional, given by

\[\begin{aligned} \partial_\vartheta \Delta_T(X_\epsilon, \vartheta, \Sigma, V) = &\frac{\beta^2 \Sigma}{T \left( 1 + \beta^2 \frac{\Sigma}{\vartheta} \right)^{5/2} \vartheta^3} \int_0^T \left[ \left( X_\epsilon(t)^2 \beta^2 - 1 \right)\vartheta - \beta^2 \Sigma \right] \exp\left( -\frac{\beta^2 X_\epsilon(t)^2}{2 (1 + \beta^2 \frac{\Sigma}{\vartheta})} \right) \, dt \\[0.25cm] &+ \frac{\beta^2 \Sigma}{\left( 1 + 2 \beta^2 \frac{\Sigma}{\vartheta} \right)^{3/2} \vartheta^2}. \end{aligned}\]

is implemented. Here, $X_ϵ$ is a one-dimensional time series of length $T$, obtained from a multiscale SDE, and $\beta$ comes from k. The potential is here $V(x) = x^2/2$.

Arguments

data::Vector{Real}: one-dimensional time series $X_ϵ$.

ϑ::Real: drift coefficient $\vartheta$.

Σ::Real: positive diffusion coefficient $\Sigma$.

Examples

julia> using MDE_project
julia> data = Langevin_ϵ(1.0, func_config=LDO(), α=2.0, σ=1.0, ϵ=0.1, T=100)[1]
julia> Δ_Gaussian1D_grad(data, 1, 1)

See also Δ_Gaussian1D.

Index

MDE_project.k

MDE_project.Δ

MDE_project.Δ_Gaussian1D

MDE_project.Δ_Gaussian1D_grad

MDE_project.Δ_Gaussian2D

MDE_project.Δ_grad_Σ

MDE_project.Δ_grad_ϑ