Optimization Task of the MDE

Introduction

The MDE is based on the following minimization task

\[\begin{aligned} \hat{\vartheta}_T(X_\epsilon) := \argmin_{\vartheta \in \Theta} \Delta_T(\vartheta, X_\epsilon), \end{aligned}\]

where

\[\begin{aligned} \Delta_T(\vartheta, X_\epsilon) := \int_{\R^d} \left| \frac1T \int_0^T \exp\left(i u^\top X_\epsilon(t)\right) \, dt - \int_{\R^d} \exp\left(i u^\top x\right) \mu(\vartheta, x) \, dx \right|^2 \varphi(u) \, du. \end{aligned}\]

$\vartheta \in \Theta \subset \mathbb{R}^p$ is the parameter to be estimated and the data $X_\epsilon$ comes in form of a time series from a multiscale process, cf. Multiscale System and Homogenized Limit Pairs.

The specific implementation of the MDE depends highly on the chosen limit model, i.e. the homogenized limit process $X$ to which the multiscale process $X_\epsilon$ converges weakly as $\epsilon \rightarrow 0$. This limit model is characterized by the invariant density $\mu$ of the limit process $X$. Building an estimator on basis of the invariant density leads to parameter identification problems which is why the implementation demands a prior estimation parameter (either diffusion or drift parameter) as input.

Another component of the MDE is the weight function $\varphi$ which is, in all considered cases, chosen as a centered Gaussian density with covariance matrix $\beta^2 I_d, \beta > 0$, because this particular choice simplifies the computational cost of the implementation.

Functions

MDE_project.MDE — Function

MDE(data::Array{Real, 1}, limit_model::String, prior_parameter::Real, ϑ_initial::Float64; verbose = false)

Return MDE value for given data in form of a time series, a defining limit_model, a prior estimation parameter prior_parameter, and an initial point ϑ_initial of the involved optimization procedure.

The optimization task

\[\begin{aligned} \argmin_{\vartheta \in \Theta} \Delta_T(X_\epsilon, \vartheta, \Sigma, V), \end{aligned}\]

is implemented and solved with the Julia package Optim.jl. Here, $X_ϵ$ is a one-dimensional time series of length $T$, obtained from a multiscale SDE, $\Delta_T$ is the associated cost functional of the MDE, see Δ_Gaussian1D, $\Sigma$ is the prior_parameter, and $V$ is a potential that is, in the given case, quadratic, i.e. $V(x)=x^2/2$.

Arguments

data::Array{Real, 1}: one-dimensional time series $X_ϵ$.

limit_model::String: defining limit model; thus far only supports "Langevin" and "Fast Chaotic Noise".

prior_parameter::Real: prior estimation parameter; limit diffusion parameter in the "Langevin" case and limit drift parameter in the "Fast Chaotic Noise" case.

ϑ_initial::Float64: initial point of the numerical optimization procedure.

'verbose::Bool=false': if verbose = true, then detailed information on the optimization will be printed in real-time.

Examples

julia> using MDE_project
julia> limit_drift_parameter = 1.0
julia> data = Fast_chaotic_ϵ([1.0, 1.0, 1.0, 1.0], A=-limit_drift_parameter, B=0.0, λ=2/45, ϵ=0.1, T=1000)
julia> MDE(data, "Fast Chaotic Noise", limit_drift_parameter, 10.0)

See also Δ_Gaussian1D_grad.

MDE(data::Array{Real, 1}, limit_model::String, V::Function, prior_parameter::Real, ϑ_initial::Float64; verbose = false)

Return MDE value for given data in form of a time series, a defining limit_model, a general potential V, a prior estimation parameter prior_parameter, and an initial point ϑ_initial of the involved optimization procedure.

The optimization task

\[\begin{aligned} \argmin_{\vartheta \in \Theta} \Delta_T(X_\epsilon, \vartheta, \Sigma, V), \end{aligned}\]

is implemented and solved with the Julia package Optim.jl. Here, $X_ϵ$ is a one-dimensional time series of length $T$, obtained from a multiscale SDE, $\Delta_T$ is the associated cost functional of the MDE, see Δ, $\Sigma$ is the prior_parameter, and $V$ is a general potential.

Arguments

data::Array{Real, 1}: one-dimensional time series $X_ϵ$.

limit_model::String: defining limit model; thus far only supports "Langevin" and "Fast Chaotic Noise".

V::Function: potential $V$ that defines the invariant density of the limit model.

prior_parameter::Real: prior estimation parameter; limit diffusion parameter in the "Langevin" case and limit drift parameter in the "Fast Chaotic Noise" case.

ϑ_initial::Float64: initial point of the numerical optimization procedure.

'verbose::Bool=false': if verbose = true, then detailed information on the optimization will be printed in real-time.

Examples

$ julia --threads 10 --project=. # start julia with 10 threads and activate project

julia> using MDE_project
julia> limit_drift_parameter = 1.0
julia> data = Fast_chaotic_ϵ([1.0, 1.0, 1.0, 1.0], A=limit_drift_parameter, B=1.0, λ=2/45, ϵ=10^(-3/2), T=100)
julia> V = NLDO()[1]
julia> MDE(data, "Fast Chaotic Noise", V, limit_drift_parameter, 0.8)

See also Δ_grad_ϑ, Δ_grad_Σ, μ.

MDE(data::Array{Real, 2}, limit_diffusion::Array{Real, 2}, ϑ_initial::Array{Real, 2}; verbose = false)

Return MDE value for given data in form of a time series, a prior estimation parameter limit_diffusion, and an initial point ϑ_initial of the involved optimization procedure.

The optimization task

\[\begin{aligned} \argmin_{\vartheta \in \Theta} \Delta_T(X_\epsilon, \vartheta, \Sigma, V), \end{aligned}\]

is implemented and solved with the Julia package Optim.jl. Here, $X_ϵ$ is a two-dimensional time series of length $T$, obtained from a multiscale SDE, $\Delta_T$ is the associated cost functional of the MDE, see Δ_Gaussian2D, $\Sigma$ is the prior 2x2 limit diffusion matrix. The initial point ϑ_initial must satisfy certain parameter constraints since the optimization has nonlinear contraints, see manuscript or source code.

Arguments

data::::Array{Real, 2}: two-dimensional time series $X_ϵ$.

limit_diffusion::Array{Real, 2}: positive definite limit diffusion matrix $\Sigma \in \mathbb{R}^{2 \times 2}$.

ϑ_initial::Array{Real, 2}: initial point $\vartheta_0 \in \mathbb{R}^{2 \times 2}$ of the numerical optimization procedure.

'verbose::Bool=false': if verbose = true, then detailed information on the optimization will be printed in real-time.

Examples

julia> using MDE_project
julia> M = [4 2;2 3]
julia> σ = 1.5               
julia> p1, p2 = (x -> sin(x), x -> 1/2*sin(x))     
julia> p1_prime, p2_prime = (x-> cos(x), x -> 1/2*cos(x))
julia> CorrK = [K(p1, σ) 0 ; 0 K(p2, σ)]
julia> A = CorrK*M # true parameter; the parameter that ought to be estimated
julia> Σ = σ*CorrK
julia> data = Langevin_ϵ_2D([-5.0, -5.0], func_config=(p1_prime, p2_prime), M=M, σ=σ, ϵ=0.1, T=1000)[1]
julia> ϑ_initial = [3.0 0.5*Σ[1]; 0.5*Σ[4] 6.0]
julia> MDE(data, Σ, ϑ_initial)

See also Δ_Gaussian2D.

Index

MDE_project.MDE