Welcome to the documentation page for "MDE for Multiscale"!

MDE_project.Fast_OU_ϵ — Function

Fast_OU_ϵ(x0, y0; <keyword arguments>)

Return a two-dimensional fast-slow Ornstein-Uhlenbeck process starting at (x0, y0) as a discretized time series.

The corresponding stochastic differential equation is for $t \in [0,T]$

\[\begin{aligned} dX_ϵ(t) &= \left( \frac{1}{ϵ} σ(X_ϵ(t)) Y_ϵ(t) + h(X_ϵ(t), Y_ϵ(t)) - σ'(X_ϵ(t))σ(X_ϵ(t)) \right) dt, \quad &X_ϵ(0) = x_0, \\ dY_ϵ(t) &= -\frac{1}{ϵ^2} Y_ϵ(t) + \frac{\sqrt{2}}{ϵ} dV(t), \quad &Y_ϵ(0) = y_0. \end{aligned}\]

Here, $σ'$ is the first derivative of $σ$. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

x0::Real: initial point $x_0$ of slow process $X_ϵ$.

y0::Real: initial point $y_0$ of fast process $Y_ϵ$.

func_config: collection of the functions $h, σ$ and $σ'$ as a tuple.

ϵ::Real=0.1: small scale parameter $ϵ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

Examples

julia> h = x -> -x
julia> σ = x -> sqrt(2)
julia> σ_prime = x -> 0
julia> Fast_OU_ϵ(1.0, 1.0, func_config=(h, σ, σ_prime))

See also Fast_OU_∞, LDA, NLDAM, NSDP.

MDE_project.Fast_OU_∞ — Function

Fast_OU_∞(X0; <keyword arguments>)

Return a one-dimensional limit process, homogenized from the fast-slow Ornstein-Uhlenbeck process, starting at X0 as a discretized time series.

The corresponding stochastic differential equation is for $t \in [0,T]$

\[\begin{aligned} dX(t) = \bar{h}(X(t)) dt - \sqrt{2 σ(X(t))^2} dW(t), \quad X(0) = X_0. \end{aligned}\]

Here, $\bar{h}$ is the average of $h$ with respect to the invariant measure of the fast process $Y_ϵ$ coming from Fast_OU_ϵ. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

X0::Real: initial point $X_0$ of limit process $X$.

func_config: collection of the functions $\bar{h}$ and $σ$ as a tuple.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

Examples

julia> h_aver = x -> -x   # corresponds to an ordinary Ornstein-Uhlenbeck process
julia> σ = x -> sqrt(2)
julia> Fast_OU_∞(1.0, func_config=(h_aver, σ))

See also Fast_OU_ϵ, LDA, NLDAM, NSDP.

MDE_project.LDA — Function

LDA(A, σ)

Return a tuple of functions used for the definition of Fast_OU_ϵ.

The returned functions are

\[\begin{aligned} h(x) &= -Ax, \quad &A > 0, \\ σ(x) &= \sqrt{σ}, \quad &σ>0, \\ σ'(x) &= 0. \end{aligned}\]

They yield a linear drift with additive noise.

Arguments

A::Real=1: non-negative real number.

σ::Real=1: positive real number.

See also Fast_OU_ϵ.

MDE_project.NLDAM — Function

NLDAM(A, B, σ_a, σ_b)

Return a tuple of functions used for the definition of Fast_OU_ϵ.

The returned functions are

\[\begin{aligned} h(x) &= Ax - Bx^3, \quad &A, B > 0, \\ σ(x) &= \sqrt{σ_a + σ_b x^2}, \quad &σ_a, σ_b > 0, \\ σ'(x) &= \frac{σ_b x}{\sqrt{σ_a + σ_b x^2}}. \end{aligned}\]

They yield a nonlinear drift with additive and multiplicative noise.

Arguments

A::Real=1: non-negative real number.

B::Real=1: non-negative real number.

σ_a::Real=1: positive real number.

σ_b::Real=1: positive real number.

See also Fast_OU_ϵ.

MDE_project.NSDP — Function

NSDP(A, B, C, σ_a, σ_b)

Return a tuple of functions used for the definition of Fast_OU_ϵ.

The returned functions are

\[\begin{aligned} h(x) &= Ax + Bx^2 - Cx^3, \quad &A, B, C > 0, \\ σ(x) &= \sqrt{σ_a + σ_b x^2}, \quad &σ_a, σ_b > 0, \\ σ'(x) &= \frac{σ_b x}{\sqrt{σ_a + σ_b x^2}}. \end{aligned}\]

They yield a nonlinear, non-symmetric double-well potential drift with additive and multiplicative noise.

Arguments

A::Real=1: non-negative real number.

B::Real=1: non-negative real number.

C::Real=1: non-negative real number.

σ_a::Real=1: positive real number.

σ_b::Real=1: positive real number.

See also Fast_OU_ϵ.

MDE_project.Overdamped_LO_ϵ — Function

Overdamped_LO_ϵ(x0; <keyword arguments>)

Return a two-dimensional overdamped Langevin process with a large-scale potential and a fast oscillating part starting at (x0, y0/ϵ) as a discretized time series.

The corresponding stochastic differential equation is for $t \in [0,T]$

\[\begin{aligned} dX_ϵ(t) = -α V'(X_ϵ(t)) - \frac{1}{ϵ} p'\left( \frac{X_ϵ(t)}{ϵ} \right) dt + \sqrt{2 σ} dW(t), \quad &X_ϵ(0) = x_0, \\ dY_ϵ(t) = -\frac{α}{ϵ} V'(X_ϵ(t)) - \frac{1}{ϵ^2} p'\left( Y_ϵ(t) \right) dt + \sqrt{\frac{2 σ}{ϵ^2}} dW(t), \quad &Y_ϵ(0) = y_0. \end{aligned}\]

Here, $V$ is a large-scale potential and $p$ a $2π$-periodic function, see LDO or NLDO. Note that $Y_ϵ = X_ϵ/ϵ$. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

x0::Real: initial point $x_0$ of slow process $X_ϵ$.

func_config: collection of the functions $V, V', p$ and $p'$ as a tuple.

α::Real: non-negative drift parameter $α$.

σ::Real: positive diffusion parameter $σ$.

ϵ::Real=0.1: small scale parameter $ϵ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

See also Overdamped_LO_∞, LDO, NLDO.

MDE_project.K — Function

K(p, σ)

Return corrective constant of the cell problem of the homogenization in the overdamped Langevin case.

Arguments

p: $2π$-periodic function.

σ: positive diffusion parameter of slow process $X_ϵ$.

See also Overdamped_LO_∞.

MDE_project.Overdamped_LO_∞ — Function

Overdamped_LO_∞(X0; <keyword arguments>)

Return a one-dimensional overdamped limit Langevin process, homogenized from the multiscale overdamped Langevin process, starting at X0 as a discretized time series.

The corresponding stochastic differential equation is for $t \in [0,T]$

\[\begin{aligned} dX(t) = -α K V'(X(t)) dt + \sqrt{2 σ K} dW(t), \quad &X(0) = X_0. \end{aligned}\]

Here, $K$ is a corrective constant that comes from the cell problem of the homogenization, see alsoK, and is computed inside the function. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

X0::Real: initial point $X_0$ of slow process $X_ϵ$.

func_config: collection of the functions $V, V', p$ and $p'$ as a tuple.

α::Real: non-negative drift parameter $α$.

σ::Real: positive diffusion parameter $σ$.

ϵ::Real=0.1: small scale parameter $ϵ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

See also Overdamped_LO_ϵ, LDO, NLDO.

MDE_project.LDO — Function

LDO()

Return a tuple of functions used for the definition of Overdamped_LO_ϵ and Overdamped_LO_∞.

The returned functions are

\[\begin{aligned} V(x) &= \frac12 x^2, \\ V'(x) &= x, \\ p(x) &= sin(x), \\ p'(x) &= cos(x). \end{aligned}\]

They yield a quadratic potential drift with a sine oscillation.

See also Overdamped_LO_ϵ, Overdamped_LO_∞.

MDE_project.NLDO — Function

NLDO()

Return a tuple of functions used for the definition of Overdamped_LO_ϵ and Overdamped_LO_∞.

The returned functions are

\[\begin{aligned} V(x) &= \frac14 x^4 - \frac12 x^2, \\ V'(x) &= x^3 - x, \\ p(x) &= sin(x), \\ p'(x) &= cos(x). \end{aligned}\]

They yield a bistable potential drift with a sine oscillation.

See also Overdamped_LO_ϵ, Overdamped_LO_∞.