Multiscale System and Homogenized Limit Pairs

The considered multiscale examples fit into the setting of homogenization for SDEs where there is a system of SDEs

\[\begin{aligned} &d X_\epsilon = \left[ \frac{1}{\epsilon} a_0(X_\epsilon, Y_\epsilon) + a_1(X_\epsilon, Y_\epsilon) \right] dt + \alpha_0(X_\epsilon, Y_\epsilon) dU_t + \alpha_1(X_\epsilon, Y_\epsilon) dV_t, \qquad &\textit{\footnotesize "slow" dynamics} \\[0.25cm] &d Y_\epsilon = \left[ \frac{1}{\epsilon^2} b_0(X_\epsilon, Y_\epsilon) + \frac{1}{\epsilon} b_1(X_\epsilon, Y_\epsilon) + b_2(X_\epsilon, Y_\epsilon) \right] dt + \frac{1}{\epsilon} \beta(X_\epsilon, Y_\epsilon) dV_t, \qquad &\textit{\footnotesize "fast" dynamics} \end{aligned}\]

depending on a small scale parameter $\epsilon > 0$ and the slow process $X_\epsilon$ converges weakly, in the sense of the induced probability measures, to the solution $X$ of the homogenized limit SDE

\[\begin{aligned} dX = f(X)dt + g(X)dW_t. \end{aligned}\]

The majority of the following functions give realizations of the multiscale processes and their respective homogenized limit in form of a time series.

Functions

MDE_project.Fast_OU_ϵ — Function

Fast_OU_ϵ(x0, y0; <keyword arguments>)

Return a 2-dimensional fast-slow Ornstein-Uhlenbeck process starting at (x0, y0) as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX_ϵ(t) &= \left( \frac{1}{ϵ} σ(X_ϵ(t)) Y_ϵ(t) + h(X_ϵ(t), Y_ϵ(t)) - σ'(X_ϵ(t))σ(X_ϵ(t)) \right) dt, \quad &X_ϵ(0) = x_0, \\ dY_ϵ(t) &= -\frac{1}{ϵ^2} Y_ϵ(t) + \frac{\sqrt{2}}{ϵ} dV(t), \quad &Y_ϵ(0) = y_0. \end{aligned}\]

Here, $σ'$ is the first derivative of $σ$. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

x0::Real: initial point $x_0$ of slow process $X_ϵ$.

y0::Real: initial point $y_0$ of fast process $Y_ϵ$.

func_config: collection of the functions $h, σ$ and $σ'$ as a tuple.

ϵ::Real=0.1: positive small scale parameter $ϵ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

Examples

# linear drift with additive noise
T = 10.0
trajectory = Fast_OU_ϵ(1.0, 1.0, func_config=LDA(), ϵ=0.1, T=T)
fig = produce_trajectory_1D(trajectory, T)
#save("trajectory_fast_OU_process.pdf", fig)

# nonlinear drift with additive and multiplicative noise
T = 10.0
trajectory = Fast_OU_ϵ(1.0, 1.0, func_config=NLDAM(), ϵ=0.1, T=T)
fig = produce_trajectory_1D(trajectory, T)

# nonlinear drift with additive and multiplicative noise, non-symmetric double-well potential
T = 10.0
trajectory = Fast_OU_ϵ(1.0, 1.0, func_config=NSDP(), ϵ=0.1, T=T)
fig = produce_trajectory_1D(trajectory, T)

See also Fast_OU_∞, LDA, NLDAM, NSDP.

MDE_project.Fast_OU_∞ — Function

Fast_OU_∞(X0; <keyword arguments>)

Return a one-dimensional limit process, homogenized from the fast-slow Ornstein-Uhlenbeck process figuring in Fast_OU_ϵ, starting at X0 as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX(t) = \bar{h}(X(t)) dt - \sqrt{2 σ(X(t))^2} dW(t), \quad X(0) = X_0. \end{aligned}\]

Here, $\bar{h}$ is the average of $h$ with respect to the invariant measure of the fast process $Y_ϵ$ coming from Fast_OU_ϵ. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

X0::Real: initial point $X_0$ of limit process $X$.

func_config: collection of the functions $\bar{h}$ and $σ$ as a tuple.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

Examples

julia> h_aver = x -> -x   # corresponds to an ordinary Ornstein-Uhlenbeck process
julia> σ = x -> sqrt(2)
julia> Fast_OU_∞(1.0, func_config=(h_aver, σ))

See also Fast_OU_ϵ, LDA, NLDAM, NSDP.

MDE_project.LDA — Function

LDA(A, σ)

Return a tuple of functions used for the definition of Fast_OU_ϵ.

The returned functions are

\[\begin{aligned} &h(x) = -Ax, \quad &A > 0, \\ &σ(x) = \sqrt{σ},\quad σ'(x) = 0, \quad &σ>0. \end{aligned}\]

They yield a linear drift with additive noise.

Arguments

A::Real=1: non-negative real number.

σ::Real=1: positive real number.

See also Fast_OU_ϵ.

MDE_project.NLDAM — Function

NLDAM(A, B, σ_a, σ_b)

Return a tuple of functions used for the definition of Fast_OU_ϵ.

The returned functions are

\[\begin{aligned} &h(x) = Ax - Bx^3, \quad &A, B > 0, \\ &σ(x) = \sqrt{σ_a + σ_b x^2}, \quad σ'(x) = \frac{σ_b x}{\sqrt{σ_a + σ_b x^2}}, \quad &σ_a, σ_b > 0. \end{aligned}\]

They yield a nonlinear drift with additive and multiplicative noise.

Arguments

A::Real=2: non-negative real number.

B::Real=10: non-negative real number.

σ_a::Real=1: positive real number.

σ_b::Real=1: positive real number.

See also Fast_OU_ϵ.

MDE_project.NSDP — Function

NSDP(A, B, C, σ_a, σ_b)

Return a tuple of functions used for the definition of Fast_OU_ϵ.

The returned functions are

\[\begin{aligned} &h(x) = Ax + Bx^2 - Cx^3, \quad &A, B, C > 0, \\ &σ(x) = \sqrt{σ_a + σ_b x^2}, \quad σ'(x) = \frac{σ_b x}{\sqrt{σ_a + σ_b x^2}}, \quad &σ_a, σ_b > 0. \end{aligned}\]

They yield a nonlinear, non-symmetric double-well potential drift with additive and multiplicative noise.

Arguments

A::Real=1: non-negative real number.

B::Real=2: non-negative real number.

C::Real=5: non-negative real number.

σ_a::Real=1: positive real number.

σ_b::Real=1: positive real number.

See also Fast_OU_ϵ.

MDE_project.Langevin_ϵ — Function

Langevin_ϵ(x0; <keyword arguments>)

Return a 2-dimensional overdamped Langevin process with a large-scale potential and a fast oscillating part starting at (x0, y0/ϵ) as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX_ϵ(t) = -α V'(X_ϵ(t)) - \frac{1}{ϵ} p'\left( \frac{X_ϵ(t)}{ϵ} \right) dt + \sqrt{2 σ} dU(t), \quad &X_ϵ(0) = x_0, \\ dY_ϵ(t) = -\frac{α}{ϵ} V'(X_ϵ(t)) - \frac{1}{ϵ^2} p'\left( Y_ϵ(t) \right) dt + \sqrt{\frac{2 σ}{ϵ^2}} dU(t), \quad &Y_ϵ(0) = y_0. \end{aligned}\]

Here, $V$ is a large-scale potential and $p$ a $2π$-periodic function, see LDO or NLDO. Note that $Y_ϵ = X_ϵ/ϵ$. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

x0::Real: initial point $x_0$ of slow process $X_ϵ$.

func_config: collection of the functions $V, V', p$ and $p'$ as a tuple.

α::Real: non-negative drift parameter $α$.

σ::Real: positive diffusion parameter $σ$.

ϵ::Real=0.1: positive small scale parameter $ϵ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

Examples

# quadratic potential V with sine oscillation p
T = 10.0
trajectory = Langevin_ϵ(1.0, func_config=LDO(), α=2.0, σ=1.0, ϵ=0.1, T=T)
fig = produce_trajectory_1D(trajectory, T)
#save("trajectory_Langevin_process.pdf", fig)

See also Langevin_∞, LDO, NLDO.

MDE_project.K — Function

K(p, σ)

Return corrective constant of the cell problem of the homogenization in the overdamped Langevin case.

Arguments

p: $2π$-periodic function.

σ: positive diffusion parameter of slow process $X_ϵ$.

See also Langevin_∞.

MDE_project.Langevin_∞ — Function

Langevin_∞(X0; <keyword arguments>)

Return a one-dimensional overdamped limit Langevin process, homogenized from the multiscale overdamped Langevin process figuring in Langevin_ϵ, starting at X0 as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX(t) = -α K V'(X(t)) dt + \sqrt{2 σ K} dW(t), \quad &X(0) = X_0. \end{aligned}\]

Here, $K$ is a corrective constant that comes from the cell problem of the homogenization, see also K, and is computed inside the function. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

X0::Real: initial point $X_0$ of limit process $X$.

func_config: collection of the functions $V, V', p$ and $p'$ as a tuple.

α::Real: non-negative drift parameter $α$.

σ::Real: positive diffusion parameter $σ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

See also Langevin_ϵ, LDO, NLDO.

MDE_project.LDO — Function

LDO()

Return a tuple of functions used for the definition of Langevin_ϵ and Langevin_∞.

The returned functions are

\[\begin{aligned} &V(x) = \frac12 x^2, \quad &V'(x) = x, \\ &p(x) = \sin(x), \quad &p'(x) = \cos(x). \end{aligned}\]

They yield a quadratic potential drift with a sine oscillation.

See also Langevin_ϵ, Langevin_∞.

MDE_project.NLDO — Function

NLDO()

Return a tuple of functions used for the definition of Langevin_ϵ and Langevin_∞.

The returned functions are

\[\begin{aligned} &V(x) = \frac14 x^4 - \frac12 x^2, \quad &V'(x) = x^3 - x, \\ &p(x) = \sin(x), \quad &p'(x) = \cos(x). \end{aligned}\]

They yield a bistable potential drift with a sine oscillation.

Examples

# a slightly different potential (for illustrative reasons)
ϵ = 0.1
V(x) = -x^2 + x^4/12
p(x) = sin(x/ϵ)
x_range = range(-4,4,2000)

# create and adjust figure components; using CairoMakie.jl here
drift_fig = Figure(size=(3840,2160), fontsize = 50)
drift_ax = Axis(drift_fig[1, 1],
  # x-axis
  xlabel = L"x",
  xticks = LinearTicks(5),
  # y-axis
  yticks = LinearTicks(5),
)
Makie.xlims!(drift_ax, x_range[1], x_range[end])
colsize!(drift_fig.layout, 1, Aspect(1, 1.8))
  

V_line = lines!(drift_ax, x_range, map(V, x_range), linewidth = 10.0, color = (:darkgrey, 1.0), linestyle = :dash)
Vp_line = lines!(drift_ax, x_range, map(x->V(x)+p(x), x_range), linewidth = 3.0, color = (:black, 1.0))

axislegend(drift_ax,
[V_line, Vp_line],
[L"$x^4/12-x^2$", L"$x^4/12 - x^2 + \sin(x/%$ϵ)$"],
labelsize = 80
)

drift_fig

See also Langevin_ϵ, Langevin_∞.

MDE_project.Langevin_ϵ_2D — Function

Langevin_ϵ_2D(x0; <keyword arguments>)

Return a 4-dimensional overdamped Langevin process with a quadratic potential and a fast separable oscillating part starting at (x0, y0/ϵ) as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX_ϵ(t) &= \begin{pmatrix} dX^{(1)}_ϵ(t) \\[0.1cm] dX^{(2)}_ϵ(t) \end{pmatrix} = - M X_ϵ(t) - \frac{1}{ϵ} \begin{pmatrix} p_1'\left(X^{(1)}_ϵ(t)/ϵ\right) \\[0.1cm] p_2'\left(X^{(2)}_ϵ(t)/ϵ\right) \end{pmatrix} dt + \sqrt{2 σ} dU(t), \quad &X_ϵ(0) = x_0, \\ dY_ϵ(t) &= \frac{X_ϵ}{ϵ}. \end{aligned}\]

A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

x0::Vector{Real}: initial point $x_0 \in \mathbb{R}^{2}$ of slow process $X_ϵ$.

func_config: collection of the $2\pi$-periodic functions $p_1'$ and $p_2'$ as a tuple.

M::Array{Real}: positive definite drift matrix $M \in \mathbb{R}^{2 \times 2}$.

σ::Real: positive diffusion parameter $σ$.

ϵ::Real=0.1: positive small scale parameter $ϵ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

Examples

# quadratic potential V and fast separable oscillating part in 2D
trajectory = Langevin_ϵ_2D([-5.0, -5.0], func_config=(x-> cos(x), x -> 1/2*cos(x)), M=[4 2;2 3], σ=5.0, ϵ=0.05, T=10.0)
fig = produce_trajectory_2D(trajectory)
#save("trajectory_Langevin_process_2D.pdf", fig)

See also Langevin_∞_2D.

MDE_project.Langevin_∞_2D — Function

Langevin_∞_2D(X0; <keyword arguments>)

Return a 2-dimensional overdamped limit Langevin process, homogenized from the multiscale overdamped Langevin process figuring in Langevin_ϵ_2D, starting at X0 as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX(t) = - K M X(t) dt + \sqrt{2 σ K} dW(t), \quad &X(0) = X_0. \end{aligned}\]

Here, $K \in \mathbb{R}^{2 \times 2}$ is a corrective constant that comes from the cell problem of the homogenization and is computed inside the function. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

X0::Vector{Real}: initial point $X_0 \in \mathbb{R}^{2}$ of limit process $X$.

func_config: collection of the $2\pi$-periodic functions $p_1'$ and $p_2'$ as a tuple.

M::Array{Real}: positive definite drift matrix $M \in \mathbb{R}^{2 \times 2}$.

σ::Real: positive diffusion parameter $σ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

See also Langevin_ϵ_2D.

MDE_project.Burger_ϵ — Function

Burger_ϵ(x0, y0, z0; <keyword arguments>)

Return a three-dimensional process described through a truncated Burger's equation starting at (x0, y0, z0) as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX_ϵ(t) &= \left( ν X_ϵ(t) - \frac{1}{2ϵ} (X_ϵ(t)Y_ϵ(t) + Y_ϵ(t)Z_ϵ(t)) \right) dt, \quad &X_ϵ(0) = x_0, \\ dY_ϵ(t) &= \left( ν Y_ϵ(t) - \frac{3}{ϵ^2} Y_ϵ(t) - \frac{1}{2ϵ} (2 X_ϵ(t)Z_ϵ(t) - X_ϵ(t)^2) \right) dt + \frac{q_1}{ϵ} dV_1(t), \quad &Y_ϵ(0) = y_0, \\ dZ_ϵ(t) &= \left( ν Z_ϵ(t) - \frac{8}{ϵ^2} Y_ϵ(t) - \frac{3}{2ϵ} X_ϵ(t)Y_ϵ(t) \right) dt + \frac{q_2}{ϵ} dV_2(t), \quad &Z_ϵ(0) = z_0. \end{aligned}\]

A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

x0::Real: initial point $x_0$ of slow process $X_ϵ$.

y0::Real: initial point $y_0$ of fast process $Y_ϵ$.

z0::Real: initial point $z_0$ of fast process $Z_ϵ$.

ν::Real: positive parameter $ν$.

q1::Real: positive parameter $q_1$.

q2::Real: positive parameter $q_2$.

ϵ::Real=0.1: positive small scale parameter $ϵ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

See also Burger_∞.

MDE_project.Burger_∞ — Function

Burger_∞(X0; <keyword arguments>)

Return a one-dimensional process described through a limit truncated Burger's equation, homogenized from the multiscale truncated Burger's equation figuring in Burger_ϵ, starting at X0 as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX(t) = \left(AX(t) - BX(t)^3\right) dt + \sqrt{σ_a + σ_bX(t)^2} dW(t), \quad X(0) = X_0, \end{aligned}\]

with the paraters

\[\begin{aligned} A = ν + \frac{q_1^2}{396} + \frac{q_2^2}{352}, \quad B = \frac{1}{12}, \quad σ_a = \frac{q_1^2 q_2^2}{2112}, \quad σ_b = \frac{q_1^2}{36}. \end{aligned}\]

A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

X0::Real: initial point $X_0$ of limit process $X$.

ν::Real: positive parameter $ν$.

q1::Real: positive parameter $q_1$.

q2::Real: positive parameter $q_2$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

See also Burger_ϵ.

MDE_project.Fast_chaotic_ϵ — Function

Fast_chaotic_ϵ(xy0; <keyword arguments>)

Return the 4-dimensional solution path of a fast chaotic noise system starting at xy0 as a discretized time series.

The corresponding ordinary differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} \frac{dX_\epsilon}{dt} &= AX - BX^3 + \frac{\lambda}{\epsilon} Y_\epsilon^{(2)}, \quad &X_ϵ(0) = x_0, \\ \frac{dY_\epsilon^{(1)}}{dt} &= \frac{10}{\epsilon^2} \left( Y_\epsilon^{(2)} - Y_\epsilon^{(1)} \right), \quad &Y_\epsilon^{(1)}(0) = y_0^{(1)}, \\ \frac{dY_\epsilon^{(2)}}{dt} &= \frac{1}{\epsilon^2} \left( 28 Y_\epsilon^{(1)} - Y_\epsilon^{(2)} - Y_\epsilon^{(1)}Y_\epsilon^{(3)} \right), \quad &Y_\epsilon^{(2)}(0) = y_0^{(2)}, \\ \frac{dY_\epsilon^{(3)}}{dt} &= \frac{1}{\epsilon^2} \left( Y_\epsilon^{(1)}Y_\epsilon^{(2)} - \frac83 Y_\epsilon^{(3)} \right), \quad &Y_\epsilon^{(3)}(0) = y_0^{(3)}. \end{aligned}\]

The ODE is solved with a fourth order Runge-Kutta scheme of the DifferentialEquations.jl package.

Arguments

xy0::Vector{Real}: initial point $(x_0, y_0^{(1)}, y_0^{(2)}, y_0^{(3)}) \in \mathbb{R}^{4}$.

A::Real: positive parameter $A$.

B::Real: positive parameter $B$.

λ::Real: positive parameter $λ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the ODE solver.

See also Fast_chaotic_∞.

MDE_project.Fast_chaotic_∞ — Function

Fast_chaotic_∞(X0; <keyword arguments>)

Return a one-dimensional process described through a limit fast chaotic noise, homogenized from the multiscale fast chaotic noise system figuring in Fast_chaotic_ϵ, starting at X0 as a discretized time series.

The corresponding stochastic differential equation is defined for $t \in [0,T]$ as

\[\begin{aligned} dX(t) = \left(AX(t) - BX(t)^3\right) dt + \sqrt{σ} dW(t), \quad X(0) = X_0, \end{aligned}\]

The diffusion parameter $σ$ is given by the Green-Kubo formula

\[\begin{aligned} \sigma = \frac{\lambda^2}{2} \int_0^\infty \lim_{T \rightarrow \infty} \frac1T \int_0^T Y_{ϵ=1}^{(2)}(s) Y_{ϵ=1}^{(2)}(s+t) \, ds \, dt, \end{aligned}\]

which requires to be numerically computed or estimated through data. A simple Euler-Maruyama discretization is implemented for the generation of the time series.

Arguments

X0::Real: initial point $X_0$ of limit process $X$.

A::Real: positive drift parameter $A$.

B::Real: positive drift parameter $B$.

σ::Real: positive diffusion parameter $σ$.

T::Real=100: time horizon of time series.

dt:Real=1e-3: time discretization step used in the Euler-Maruyama scheme.

See also Fast_chaotic_ϵ.

MDE_project.produce_trajectory_1D — Function

produce_trajectory_1D(trajectory, T)

Generate a plot of a 2-dimensional (slow + fast dimension) multiscale time series of length T using the CairoMakie.jl package.

Arguments

trajectory: 2-dimensional time series of length T

T::Real: time horizon of time series.

Examples

# quadratic potential V with sine oscillation p
T = 10.0
trajectory = Langevin_ϵ(1.0, func_config=LDO(), α=2.0, σ=1.0, ϵ=0.1, T=T)
fig = produce_trajectory_1D(trajectory, T)

See also produce_trajectory_2D.

MDE_project.produce_trajectory_2D — Function

produce_trajectory_2D(trajectory)

Generate a plot of a 4-dimensional (slow + fast dimension) multiscale time series using the CairoMakie.jl package.

Arguments

trajectory: 4-dimensional time series

Examples

# quadratic potential V and fast separable oscillating part in 2D
trajectory = Langevin_ϵ_2D([-5.0, -5.0], func_config=(x-> cos(x), x -> 1/2*cos(x)), M=[4 2;2 3], σ=5.0, ϵ=0.05, T=10.0)
fig = produce_trajectory_2D(trajectory)

See also produce_trajectory_1D.

Index

MDE_project.Burger_ϵ

MDE_project.Burger_∞

MDE_project.Fast_OU_ϵ

MDE_project.Fast_OU_∞

MDE_project.Fast_chaotic_ϵ

MDE_project.Fast_chaotic_∞

MDE_project.K

MDE_project.LDA

MDE_project.LDO

MDE_project.Langevin_ϵ

MDE_project.Langevin_ϵ_2D

MDE_project.Langevin_∞

MDE_project.Langevin_∞_2D

MDE_project.NLDAM

MDE_project.NLDO

MDE_project.NSDP

MDE_project.produce_trajectory_1D

MDE_project.produce_trajectory_2D